+++ server/app.js
@@ -1,6 +1,7 @@
 import express from 'express';
 import bodyParser from "body-parser";
-import route from './routes/api_routes';
+import user from './routes/user_routes';
+import messages from './routes/messages_routes';
 
 // init App
 const app = express();
@@ -14,10 +15,11 @@ const port = process.env.PORT || 3000;
 // body Parser middleware
 app.use(bodyParser.json());
 
-app.use('/', route);
+app.use('/', user);
+app.use('/', messages);
 
 app.listen(port, () => {
     console.log(`Strated up at port ${port}`);
 });
 
-module.exports = app;
+export default app;
