+++ server/routes/messages_routes.js
@@ -1,61 +1,16 @@
 import express from 'express';
 import {
- addUser, getUser, createMessage, getMessageById, getMessages, deleteMessageById,
+    createMessage, getMessageById, getMessages, deleteMessageById,
 } from "../utils/utils";
 
 const router = express.Router();
 
-// POST signup
-router.post('/api/v1/auth/signup', (req, res) => {
-    const {
- id, email, password, firstName, lastName, token,
-} = req.body;
-    const user = {
-        id, email, password, firstName, lastName, token,
-    };
-    addUser(user).then(() => {
-        res.status(200).send({
-            status: 200,
-            data: [{
-                token,
-            }],
-        });
-    }).catch((err) => {
-        res.status(400).send({
-            status: 400,
-            error: err,
-        });
-    });
-});
-
-// POST login
-router.post('/api/v1/auth/login', (req, res) => {
-    const { email, password } = req.body;
-    const login = {
-        email, password,
-    };
-
-    getUser(login).then((user) => {
-        const { token } = user[0];
-        res.status(200).send({
-            status: 200,
-            data: [{
-                token,
-            }],
-        });
-    }).catch((err) => {
-        res.status(400).send({
-            status: 400,
-            error: err,
-        });
-    });
-});
 
 // POST messages
 router.post('/api/v1/messages', (req, res) => {
     const {
- subject, message, id, createdOn,
-} = req.body;
+        subject, message, id, createdOn,
+    } = req.body;
     const status = "read";
     const parentMessageId = 105;
     const messag = {
@@ -172,4 +127,4 @@ router.delete('/api/v1/messages/:messageId', (req, res) => {
     });
 });
 
-module.exports = router;
+export default router;
